requirements.txt
fastapi
uvicorn
networkx
leidenalg
igraph
numpy
scipy
pandas
openai
python-dotenv
tiktoken
pydantic
pydantic-settings
